+++ ExamifyApp/ExaminationPL/Controllers/Admin/StudentController.cs
@@ -19,33 +19,65 @@ namespace Examination.PL.Controllers.Admin
         }
         public IActionResult getAll()
         {
-            var Data = _studentRepo.getAllStudent();
+            int? UserId = HttpContext.Session.GetInt32("UserId");
+            int? RoleID = HttpContext.Session.GetInt32("RoleId");
+            if (UserId != null && RoleID==1)
+            {
+                var Data = _studentRepo.getAllStudent();
             return View(Data);
+            }
+            return RedirectToAction("Login", "Account");
+
         }
         public IActionResult getById(int id)
         {
-            var Data = _studentRepo.GetStudentById(id);
+            int? UserId = HttpContext.Session.GetInt32("UserId");
+            int? RoleID = HttpContext.Session.GetInt32("RoleId");
+            if (UserId != null && RoleID==1)
+            {
+                var Data = _studentRepo.GetStudentById(id);
             return View(Data);
+            }
+            return RedirectToAction("Login", "Account");
+
         }
         public IActionResult DeleteInstructorByID(int Id)
         {
-            _studentRepo.DeleteInstructorByID(Id);
+            int? UserId = HttpContext.Session.GetInt32("UserId");
+            int? RoleID = HttpContext.Session.GetInt32("RoleId");
+            if (UserId != null && RoleID==1)
+            {
+                _studentRepo.DeleteInstructorByID(Id);
             return RedirectToAction("getAll");
+            }
+            return RedirectToAction("Login", "Account");
+
         }
 
 
         public IActionResult InsertStudent()
         {
-            ViewData["Department"] = new SelectList(departmentRepo.GetAllDepartments(), "DeptId", "DeptName");
+            int? UserId = HttpContext.Session.GetInt32("UserId");
+            int? RoleID = HttpContext.Session.GetInt32("RoleId");
+            if (UserId != null && RoleID==1)
+            {
+                ViewData["Department"] = new SelectList(departmentRepo.GetAllDepartments(), "DeptId", "DeptName");
 
             return View();
+            }
+            return RedirectToAction("Login", "Account");
+
         }
         [HttpPost]
         public IActionResult InsertStudent(InsertStudentVM insertStudentVM)
         {
-            if(ModelState.IsValid)
+            int? UserId = HttpContext.Session.GetInt32("UserId");
+            int? RoleID = HttpContext.Session.GetInt32("RoleId");
+            if (UserId != null && RoleID==1)
             {
-                insertStudentVM.StImg = FileUploader.UploadFile("Imgs", insertStudentVM.Image);
+                if (ModelState.IsValid)
+            {
+                insertStudentVM.StImg = FileUploader.UploadFile("StudentsImages", insertStudentVM.Image);
 
                 _studentRepo.InsertStudent(insertStudentVM);
                 return RedirectToAction("getAll");
@@ -53,28 +85,46 @@ namespace Examination.PL.Controllers.Admin
             ViewData["Department"] = new SelectList(departmentRepo.GetAllDepartments(), "DeptId", "DeptName");
 
             return View(insertStudentVM);
+            }
+            return RedirectToAction("Login", "Account");
+
         }
 
         [HttpGet]
         public IActionResult Edit(int id)
         {
-            ViewData["Department"] = new SelectList(departmentRepo.GetAllDepartments(), "DeptId", "DeptName");
+            int? UserId = HttpContext.Session.GetInt32("UserId");
+            int? RoleID = HttpContext.Session.GetInt32("RoleId");
+            if (UserId != null && RoleID==1)
+            {
+                ViewData["Department"] = new SelectList(departmentRepo.GetAllDepartments(), "DeptId", "DeptName");
 
             var Data = _studentRepo.GetStudentDataById(id);
             return View(Data);
+            }
+            return RedirectToAction("Login", "Account");
+
         }
         [HttpPost]
         public IActionResult Edit(EditStudentVM model)
         {
-            if(ModelState.IsValid)
+            int? UserId = HttpContext.Session.GetInt32("UserId");
+            int? RoleID = HttpContext.Session.GetInt32("RoleId");
+            if (UserId != null && RoleID==1)
             {
-                _studentRepo.Edit(model);
+                if (ModelState.IsValid)
+            {
+                    model.StImg = FileUploader.UploadFile("StudentsImages", model.Image);
+
+                    _studentRepo.Edit(model);
                 return RedirectToAction("getAll");
 
             }
 
             ViewData["Department"] = new SelectList(departmentRepo.GetAllDepartments(), "DeptId", "DeptName");
             return View(model);
+            }
+            return RedirectToAction("Login", "Account");
 
         }
     }
