+++ ExamifyApp/ExaminationPL/Controllers/CourseController.cs
@@ -15,18 +15,34 @@ namespace ExaminationPL.Controllers
         }
         public IActionResult getAll()
         {
-            var Data = courseRepo.GetAll();
+            int? UserId = HttpContext.Session.GetInt32("UserId");
+            int? RoleID = HttpContext.Session.GetInt32("RoleId");
+            if (UserId != null && RoleID==1)
+            {
+                var Data = courseRepo.GetAll();
             return View(Data);
+            }
+            return RedirectToAction("Login", "Account");
         }
         public IActionResult DeleteCourse(int id)
         {
-            courseRepo.DeleteCourse(id);
+            int? UserId = HttpContext.Session.GetInt32("UserId");
+            int? RoleID = HttpContext.Session.GetInt32("RoleId");
+            if (UserId != null && RoleID==1)
+            {
+                courseRepo.DeleteCourse(id);
             return RedirectToAction("getAll");
+            }
+            return RedirectToAction("Login", "Account");
         }
         [HttpGet]
         public IActionResult EditCourse(int id)
         {
-           var Data = courseRepo.getCourseById(id);
+            int? UserId = HttpContext.Session.GetInt32("UserId");
+            int? RoleID = HttpContext.Session.GetInt32("RoleId");
+            if (UserId != null && RoleID==1)
+            {
+                var Data = courseRepo.getCourseById(id);
             EditCourseVM editCourseVM = new EditCourseVM()
             {
                 CrsId = Data.CrsId,
@@ -34,41 +50,67 @@ namespace ExaminationPL.Controllers
                 CrsName = Data.CrsName,
             };
             return View(editCourseVM);
+            }
+            return RedirectToAction("Login", "Account");
 
         }
         [HttpPost]
         public IActionResult EditCourse(EditCourseVM editCourseVM)
         {
-            //2 
-            if(ModelState.IsValid) 
+            int? UserId = HttpContext.Session.GetInt32("UserId");
+            int? RoleID = HttpContext.Session.GetInt32("RoleId");
+            if (UserId != null && RoleID==1)
+            {
+                //2 
+                if (ModelState.IsValid) 
             {
                 courseRepo.EditCourse(editCourseVM);
                 return RedirectToAction("getAll");
             }
             return View(editCourseVM);
+            }
+            return RedirectToAction("Login", "Account");
 
         }
         public IActionResult getCourseById(int id)
         {
-            var Data = courseRepo.getCourseById(id);
+            int? UserId = HttpContext.Session.GetInt32("UserId");
+            int? RoleID = HttpContext.Session.GetInt32("RoleId");
+            if (UserId != null && RoleID==1)
+            {
+                var Data = courseRepo.getCourseById(id);
             return View(Data);
+            }
+            return RedirectToAction("Login", "Account");
         }
 
         [HttpGet]
         public IActionResult InsertCourse(int id)
         {
-            return View();
+            int? UserId = HttpContext.Session.GetInt32("UserId");
+            int? RoleID = HttpContext.Session.GetInt32("RoleId");
+            if (UserId != null && RoleID==1)
+            {
+                return View();
+            }
+            return RedirectToAction("Login", "Account");
         }
         [HttpPost]
         public IActionResult InsertCourse(InsertCourseVM model)
         {
-            //2 
-            if (ModelState.IsValid)
+            int? UserId = HttpContext.Session.GetInt32("UserId");
+            int? RoleID = HttpContext.Session.GetInt32("RoleId");
+            if (UserId != null && RoleID==1)
+            {
+                //2 
+                if (ModelState.IsValid)
             {
                 courseRepo.InsertCourse(model);
                 return RedirectToAction("getAll");
             }
             return View(model);
+            }
+            return RedirectToAction("Login", "Account");
 
         }
     }
